+++ php/lib/Ice.php
@@ -41,110 +41,6 @@ namespace Ice
         abstract public function ice_id();
     }
 
-    class Value
-    {
-        public $_ice_slicedData;
-
-        public static function ice_staticId()
-        {
-            return "::Ice::Object";
-        }
-
-        public function ice_id()
-        {
-            return "::Ice::Object";
-        }
-
-        public function ice_getSlicedData()
-        {
-            return $this->_ice_slicedData;
-        }
-    }
-
-    class ObjectPrxHelper
-    {
-        public static function checkedCast($proxy, $facetOrContext=null, $context=null)
-        {
-            return $proxy->ice_checkedCast('::Ice::Object', $facetOrContext, $context);
-        }
-
-        public static function uncheckedCast($proxy, $facet=null)
-        {
-            return $proxy->ice_uncheckedCast('::Ice::Object', $facet);
-        }
-
-        public static function ice_staticId()
-        {
-            return '::Ice::Object';
-        }
-    }
-
-    global $Ice__t_Value;
-    global $Ice__t_ObjectSeq;
-    global $Ice__t_ObjectPrx;
-    global $Ice__t_ObjectProxySeq;
-
-    $Ice__t_Value = IcePHP_defineClass('::Ice::Object', "\\Ice\\Value", -1, false, null, null);
-    $Ice__t_ObjectSeq = IcePHP_defineSequence('::Ice::ObjectSeq', $Ice__t_Value);
-    $Ice__t_ObjectPrx = IcePHP_defineProxy('::Ice::Object', null, null);
-    $Ice__t_ObjectProxySeq = IcePHP_defineSequence('::Ice::ObjectProxySeq', $Ice__t_ObjectPrx);
-
-    class UnknownSlicedValue extends Value
-    {
-        public $unknownTypeId;
-
-        public function __construct()
-        {
-        }
-
-        public function ice_id()
-        {
-            return $this->unknownTypeId;
-        }
-    }
-
-    global $Ice__t_UnknownSlicedValue;
-    $Ice__t_UnknownSlicedValue = IcePHP_defineClass('::Ice::UnknownSlicedValue', "\\Ice\\UnknownSlicedValue", -1, false,
-                                                    $Ice__t_Value, null);
-
-    interface ValueFactory
-    {
-        public function create($id);
-    }
-
-    class InitializationData
-    {
-        public function __construct($properties=null, $logger=null)
-        {
-            $this->properties = $properties;
-            $this->logger = $logger;
-        }
-
-        public $properties;
-        public $logger;
-    }
-
-    class SlicedData
-    {
-        public $slices;
-    }
-
-    class SliceInfo
-    {
-        public $typeId;
-        public $compactId;
-        public $bytes;
-        public $instances;
-        public $hasOptionalMembers;
-        public $isLastSlice;
-    }
-
-    class FormatType
-    {
-        const DefaultFormat = 0;
-        const CompactFormat = 1;
-        const SlicedFormat = 2;
-    }
 }
 
 namespace
@@ -152,11 +48,16 @@ namespace
     //
     // Include certain generated files.
     //
+    require_once 'Ice/Proxy.php';
+    require_once 'Ice/Value.php';
+
     require_once 'Ice/BuiltinSequences.php';
     require_once 'Ice/CompressBatch.php';
     require_once 'Ice/ConnectionClose.php';
     require_once 'Ice/EndpointSelectionType.php';
     require_once 'Ice/EndpointTypes.php';
+    require_once 'Ice/Identity.php';
+    require_once 'Ice/InitializationData.php';
     require_once 'Ice/LocalExceptions.php';
     require_once 'Ice/Locator.php';
     require_once 'Ice/Metrics.php';
@@ -181,93 +82,4 @@ namespace
     IcePHP_defineOperation($Ice__t_ObjectPrx, 'ice_ids', 2, 0, null, null, array($Ice__t_StringSeq), null);
 }
 
-namespace Ice
-{
-    $Protocol_1_0 = new ProtocolVersion(1, 0);
-    $Encoding_1_0 = new EncodingVersion(1, 0);
-    $Encoding_1_1 = new EncodingVersion(1, 1);
-
-    //
-    // Proxy comparison functions.
-    //
-    function proxyIdentityCompare($lhs, $rhs)
-    {
-        if(($lhs != null && !($lhs instanceof ObjectPrx)) || ($rhs != null && !($rhs instanceof ObjectPrx)))
-        {
-            throw new InvalidArgumentException('argument is not a proxy');
-        }
-        if($lhs == null && $rhs == null)
-        {
-            return 0;
-        }
-        elseif($lhs == null && $rhs != null)
-        {
-            return -1;
-        }
-        elseif($lhs != null && $rhs == null)
-        {
-            return 1;
-        }
-        else
-        {
-            $lid = $lhs->ice_getIdentity();
-            $rid = $rhs->ice_getIdentity();
-            if($lid < $rid)
-            {
-                return -1;
-            }
-            elseif($lid > $rid)
-            {
-                return 1;
-            }
-            else
-            {
-                return 0;
-            }
-        }
-    }
-
-    function proxyIdentityEqual($lhs, $rhs)
-    {
-        return proxyIdentityCompare($lhs, $rhs) == 0;
-    }
-
-    function proxyIdentityAndFacetCompare($lhs, $rhs)
-    {
-        $n = proxyIdentityCompare($lhs, $rhs);
-        if($n == 0 && $lhs != null && $rhs != null)
-        {
-            $n = strcmp($lhs->ice_getFacet(), $rhs->ice_getFacet());
-        }
-        return $n;
-    }
-
-    function proxyIdentityAndFacetEqual($lhs, $rhs)
-    {
-        return proxyIdentityAndFacetCompare($lhs, $rhs) == 0;
-    }
-}
-
-namespace
-{
-    function Ice_proxyIdentityCompare($lhs, $rhs)
-    {
-        return Ice\proxyIdentityCompare($lhs, $rhs);
-    }
-
-    function Ice_proxyIdentityEqual($lhs, $rhs)
-    {
-        return Ice\proxyIdentityEqual($lhs, $rhs);
-    }
-
-    function Ice_proxyIdentityAndFacetCompare($lhs, $rhs)
-    {
-        return Ice\proxyIdentityAndFacetCompare($lhs, $rhs);
-    }
-
-    function Ice_proxyIdentityAndFacetEqual($lhs, $rhs)
-    {
-        return Ice\proxyIdentityAndFacetEqual($lhs, $rhs);
-    }
-}
 ?>
